+++ target/package/fp-ts-0.0.1/target/debug/libfp_ts.d
@@ -1,1 +1,0 @@
-/home/wizard/Dev/fp-rs/target/package/fp-ts-0.0.1/target/debug/libfp_ts.rlib: /home/wizard/Dev/fp-rs/target/package/fp-ts-0.0.1/src/lib.rs