+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Feat
+
+- Add functor
+- Add functor implem for Option & Result
+
 ## [0.1.0]
 
 ### Added